Revert "Replace EnsureConsistencyX with new patch"
This reverts commit d415c4516eef6534556dd8a0f958efa9914779c2.
